No need to use the abstracted `backend.js` to launch Cypress tests
